Create test-server.key
In general don't ever ever ever share a private key like this. This is a throwaway key in a throwaway project which is why I'm doing it. Anyone who has access to an important key could impersonate an important server, or use the key to decrypt all traffic between anyone and an important server, which defeats the whole purpose of encryption. So it's really bad.